collisions + MODULES?!
Tilemap collisions work great! :D
I'm trying to work on collisions with npcs...not so great. A bit messy and errory right now, but I'm leaving it for another day. "this" is getting me. :(
Also I switched everything to ES6 modules strict mode stuffs :/
I like it but it was a pain adjusting with the import and export.